Code Update: Temper the Steel
* Improved Exception handling code
* Activated Grinch File Management System and fixed the code issues
* Fixed a few minor issues in WraithRead.java
* Integrated Setup code into the Loader code, should improve program security
* Shifted a few output streams to IOStreams in NionKernel.java
* Deleted Setup.java
